Flex mode by default
Flex mode (when the sum of a row is <= 1) is now the default mode, with no selection key.
During optimization tests, there were no difference in the results. For most of the resulting matrices their row sums were >0.99.
Thus the mode is not harmful and might be useful in future.