Global update: base, y_key, zs_c
EdgeFactory now works with a list of basic edges and adds new to that list. List can be either empy and the procedure will work as previously, or can have some edges, e.g. an efficient MST solution.

A more advanced measure of subgraph z-score characteristic is available now as zs_c. It correlates Pearson correlation between a given target vector and the z-scores of the networks subgraphs.

GeneticAlgorithm is generallized to optimize not only the edge robustness, but rather any of: motifs score, z-score correlation, node robustness, edge robustness.